rename files for BinaryCode and SymbolicBinary
