+++ src/openfermion/ops/__init__.py
@@ -10,15 +10,15 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-from ._binary_operator import SymbolicBinary
-from ._code_operator import (BinaryCode,
-                             linearize_decoder)
 from ._polynomial_tensor import PolynomialTensor, general_basis_change
 from ._quadratic_hamiltonian import QuadraticHamiltonian
+from ._symbolic_binary import SymbolicBinary
 from ._symbolic_operator import (SymbolicOperator,
                                  prune_unused_indices)
 
 # Imports out of alphabetical order to avoid circular dependency.
+from ._binary_code import (BinaryCode,
+                           linearize_decoder)
 from ._fermion_operator import (FermionOperator,
                                 freeze_orbitals,
                                 normal_ordered)
