+++ src/openfermion/ops/_binary_code.py
