+++ src/openfermion/ops/_symbolic_binary.py
